将整数换成小数，使运算符左右两边类型相同

+++ R.vert.glsl
@@ -1,5 +1,5 @@
 attribute vec3 kzPosition;
-uniform highp mat4 kzProjectionCameraWorldMatrix;
+uniform mediump mat4 kzProjectionCameraWorldMatrix;
 attribute vec2 kzTextureCoordinate0;
 
 varying vec2 vTexCoord;
@@ -9,6 +9,7 @@ void main()
     precision mediump float;
     vTexCoord = kzTextureCoordinate0;
 
-    // kzPosition等是kanzi自动映射好的。
     gl_Position = kzProjectionCameraWorldMatrix * vec4(kzPosition.xyz, 1.0);
 }
+
+
